+++ src/hooks/useFilter.ts
@@ -1,12 +1,6 @@
 import { useLocation } from "react-router-dom";
 import { SetStateAction, useEffect } from "react";
-
 import useNavigateWithParams from "./useNavigateWithParams";
-import { useAtom } from "jotai";
-import {
-  paginationAtom,
-  prevPaginationStateAtom,
-} from "@/store/paginationAtom";
 import { ParentPageType } from "@/types/PaginationType";
 import { NavigateMode, SupportedFilterTypes } from "@/types/FilterType";
 
@@ -26,23 +20,10 @@ const useFilter = <TFilter extends SupportedFilterTypes>({
   itemId,
 }: Props<TFilter>) => {
   const { search } = useLocation();
-  const { navigateWithParams } = useNavigateWithParams(parentPage);
-  const [, setPaginationState] = useAtom(paginationAtom);
-  const [, setPrevPaginationState] = useAtom(prevPaginationStateAtom);
-
+  const { navigateWithParams } = useNavigateWithParams();
   const queryParams = new URLSearchParams(search);
   const sort = queryParams.get("sort");
   const direction = queryParams.get("direction");
-
-  const resetPaginationState = () => {
-    setPaginationState((prev) => ({
-      ...prev,
-      currentPage: 1,
-      pagePosition: "START",
-      middlePages: [],
-      isMiddlePagesUpdated: false,
-    }));
-  };
 
   // URL의 쿼리 파라미터와 동기화
   useEffect(() => {
@@ -60,8 +41,7 @@ const useFilter = <TFilter extends SupportedFilterTypes>({
 
   const onOptionClick = (filter: TFilter) => {
     if (type === "queryString" && parentPage) {
-      setPrevPaginationState(undefined);
-      resetPaginationState();
+      // setPrevPaginationState(undefined);
 
       navigateWithParams({
         filter,
@@ -71,7 +51,7 @@ const useFilter = <TFilter extends SupportedFilterTypes>({
       });
     } else if (type === "state") {
       setFilterCriteria(filter as TFilter);
-      resetPaginationState();
+      // resetPaginationState();
     }
   };
   return { onOptionClick };
