+++ src/store/paginationAtom.ts
@@ -21,10 +21,14 @@ const initialPaginationState: PaginationType = {
 // 현재 쿼리스트링 페이지네이션 부분에서는 모두 같은 paginationAtom을 사용하고 있습니다.
 export const paginationAtom = atom<PaginationType>(initialPaginationState);
 
-// 이전 (책 목록) 페이지네이션 상태 저장
-export const prevPaginationStateAtom = atom<PaginationType | undefined>(
-  undefined,
-);
+// // 이전 (책 목록) 페이지네이션 상태 저장
+// export const prevPaginationStateAtom = atom<PaginationType | undefined>(
+//   undefined,
+// );
+// // 이전 (책 목록) 페이지네이션 position 상태 저장
+// export const prevPagePositionAtom = atom<PagePositionType | undefined>(
+//   undefined,
+// );
 
 // // 마이페이지 > 만든 퀴즈
 // export const myMadeQuizPaginationAtom = atom<PaginationType>({
