+++ src/types/FilterType.ts
@@ -11,7 +11,6 @@ export const SortOptions = {
   quizzes: ["STAR_RATING", "CREATED_AT", "UPDATED_AT"] as const,
   myMadeQuizzes: ["CREATED_AT", "UPDATED_AT", "TITLE"] as const,
   mySolvedQuizzes: ["TITLE", "CREATED_AT"] as const,
-  myDraftQuizzes: ["TITLE", "CREATED_AT"] as const,
   myStudyUnSolvedQuizzes: ["TITLE", "CREATED_AT"] as const,
   myStudySolvedQuizzes: ["TITLE", "CREATED_AT"] as const,
 } as const;
@@ -24,8 +23,6 @@ export type QuizzesSortType = (typeof SortOptions.quizzes)[number];
 export type MyMadeQuizzesSortType = (typeof SortOptions.myMadeQuizzes)[number];
 export type MySolvedQuizzesSortType =
   (typeof SortOptions.mySolvedQuizzes)[number];
-// export type MyDraftQuizzesSortType =
-//   (typeof SortOptions.myDraftQuizzes)[number];
 export type MyStudyUnSolvedQuizzesSortType =
   (typeof SortOptions.myStudyUnSolvedQuizzes)[number];
 export type MyStudySolvedQuizzesSortType =
@@ -44,7 +41,6 @@ export type StudyGroupsFilterType = FilterType<StudyGroupsSortType>;
 export type QuizzesFilterType = FilterType<QuizzesSortType>;
 export type MyMadeQuizzesFilterType = FilterType<MyMadeQuizzesSortType>;
 export type MySolvedQuizzesFilterType = FilterType<MySolvedQuizzesSortType>;
-// export type MyDraftQuizzesFilterType = FilterType<MyDraftQuizzesSortType>;
 export type MyStudyUnSolvedQuizzesFilterType =
   FilterType<MyStudyUnSolvedQuizzesSortType>;
 export type MyStudySolvedQuizzesFilterType =
